fixed converter for table read, added plain loader
