+++ src/layouts/LoadingLayout.tsx
@@ -1,10 +1,11 @@
-import LoadingMainLogo from "@/components/LoadingMainLogo";
+import Loading from "@/components/loaders/Loading";
+import LoadingMainLogo from "@/components/loaders/LoadingMainLogo";
 import { Center } from "@mantine/core";
 
-export default function LoadingLayout(props: {fullscreen?: boolean}) {
+export default function LoadingLayout(props: {fullscreen?: boolean, logo?: boolean}) {
   return(
     <Center className={props.fullscreen ? "min-h-screen" : "min-h-full"}>
-      <LoadingMainLogo />
+      {props.logo ? <LoadingMainLogo /> : <Loading />}
     </Center>
   );
 }
