+++ src/layouts/table/tableSelection/CreateJoinTableLayout.tsx
@@ -7,7 +7,7 @@ export default function CreateJoinTableLayout() {
   const [activeTab, setActiveTab] = useState<string | null>('Join');
 
   return (
-    <Center className="flex-col rounded-3xl shadow-red-100 shadow-xl bg-gradient-to-r from-rose-50 via-white to-rose-50 p-10 m-10">
+    <Center className="flex-col rounded-3xl bg-white shadow-red-100 shadow-xl p-10 m-10">
       <Text className="mb-10 text-5xl font-black" variant="gradient" gradient={{from: "red.7", to: "red.4"}}>Have a Seat</Text>
       <CreateJoinTableTabs activeTab={activeTab} setActiveTab={setActiveTab}/>
     </Center>
